requirements.txt
aiogram==3.11.0
python-decouple==3.8